requirements.txt
SQLModel
fastapi[standard]
alembic
python-dotenv
python-jose[cryptography] 
passlib[bcrypt] 
python-multipart
pydantic-settings